feat(generator): canonical PHP type names (bool, float) and short array syntax
Phase B task 2.2 per docs/plans/2026-02-03-builder-om-modernization.md.

Source change: PropelTypes BOOLEAN/REAL/FLOAT/DOUBLE/BOOLEAN_EMU
NATIVE_TYPE constants now resolve to canonical PHP names ('bool',
'float') instead of legacy aliases ('boolean', 'double'). The
isPhpPrimitiveType / isPhpPrimitiveNumericType helpers were extended
to accept both the canonical and legacy names so existing user code
that hand-sets phpType to 'boolean' or 'double' keeps working.
ColumnMutatorBuilderTrait emits (bool) cast and []-array literal,
ColumnAccessorBuilderTrait emits []-array literal.

Golden-diff: 10 generated classes touched. Boolean columns now show
@var bool|null / @return bool|null / (bool) cast in hydrate path;
SET-column getters use [] instead of array() in the lazy-init branch.

Suite remains green: 2407 / 5158 / 21.

+++ tests/snapshots/bookstore-golden/Propel/Tests/Bookstore/Base/Book.php
@@ -122,7 +122,7 @@ abstract class Book implements ActiveRecordInterface
     /**
      * The value for the price field.
      * Price of the book.
-     * @var        double|null
+     * @var        float|null
      */
     protected $price;
 
@@ -579,7 +579,7 @@ abstract class Book implements ActiveRecordInterface
     /**
      * Get the [price] column value.
      * Price of the book.
-     * @return double|null
+     * @return float|null
      */
     public function getPrice()
     {
@@ -669,13 +669,13 @@ abstract class Book implements ActiveRecordInterface
     /**
      * Set the value of [price] column.
      * Price of the book.
-     * @param double|null $v New value
+     * @param float|null $v New value
      * @return $this The current object (for fluent API support)
      */
     public function setPrice($v)
     {
         if ($v !== null) {
-            $v = (double) $v;
+            $v = (float) $v;
         }
 
         if ($this->price !== $v) {
@@ -780,7 +780,7 @@ abstract class Book implements ActiveRecordInterface
             $this->isbn = (null !== $col) ? (string) $col : null;
 
             $col = $row[TableMap::TYPE_NUM == $indexType ? 3 + $startcol : BookTableMap::translateFieldName('Price', TableMap::TYPE_PHPNAME, $indexType)];
-            $this->price = (null !== $col) ? (double) $col : null;
+            $this->price = (null !== $col) ? (float) $col : null;
 
             $col = $row[TableMap::TYPE_NUM == $indexType ? 4 + $startcol : BookTableMap::translateFieldName('PublisherId', TableMap::TYPE_PHPNAME, $indexType)];
             $this->publisher_id = (null !== $col) ? (int) $col : null;
